Written first catalog example tests
