--- features/10_mins_tutorial.feature
@@ -1,13 +1,0 @@
-Feature: Is it Friday yet?
-  Everybody wants to know when it's Friday
-
-  Scenario Outline: Today is or is not Friday
-    Given today is "<day>"
-    When I ask whether it's Friday yet
-    Then I should be told "<answer>"
-
-  Examples:
-    | day            | answer |
-    | Friday         | TGIF   |
-    | Sunday         | Nope   |
-    | anything else! | Nope   |
