+++ features/step_definitions/tokens.js
@@ -1,0 +1,17 @@
+const { Given } = require('@cucumber/cucumber');
+
+const TOKEN = {
+  'PagoPA S.p.A.': {
+    admin: '',
+  },
+  'Comune di Milano': {
+    security: ''
+  },
+  Sogecap: {
+    admin: ''
+  }
+}
+
+Given('l\'utente è un {string} di {string}', function(role, party) {
+  this.token = TOKEN[party][role]
+})
